requirements.txt
streamlit
pandas
plotly
openai
python-dotenv
jinja2
groq